+++ libraries/test_session_current/src/androidTest/java/androidx/media3/session/MediaControllerListenerTest.java
@@ -83,6 +83,7 @@ import androidx.media3.test.session.common.MainLooperTestRule;
 import androidx.media3.test.session.common.MediaSessionConstants;
 import androidx.media3.test.session.common.SurfaceActivity;
 import androidx.media3.test.session.common.TestUtils;
+import androidx.media3.test.utils.FakeTimeline;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
@@ -4145,6 +4146,64 @@ public class MediaControllerListenerTest {
     assertThat(latch.await(TIMEOUT_MS, MILLISECONDS)).isTrue();
     assertThat(listener1States).containsExactly(Player.STATE_READY, Player.STATE_IDLE).inOrder();
     assertThat(listener2States).containsExactly(Player.STATE_READY, Player.STATE_IDLE).inOrder();
+  }
+
+  @Test
+  public void
+      missingCommandGetTimeline_withCommandGetCurrentMediaItem_doesNotIncludePlaceholderFlag()
+          throws Exception {
+    MediaController controller = controllerTestRule.createController(remoteSession.getToken());
+    remoteSession
+        .getMockPlayer()
+        .setTimeline(
+            new FakeTimeline(
+                new FakeTimeline.TimelineWindowDefinition.Builder().setPlaceholder(true).build(),
+                new FakeTimeline.TimelineWindowDefinition.Builder().setPlaceholder(true).build()));
+    CountDownLatch timelineLatch = new CountDownLatch(1);
+    List<Timeline> capturedTimelines = new ArrayList<>();
+    Player.Listener timelineListener =
+        new Player.Listener() {
+          @Override
+          public void onTimelineChanged(Timeline timeline, int reason) {
+            capturedTimelines.add(timeline);
+            timelineLatch.countDown();
+          }
+        };
+    controller.addListener(timelineListener);
+    remoteSession
+        .getMockPlayer()
+        .notifyTimelineChanged(Player.TIMELINE_CHANGE_REASON_SOURCE_UPDATE);
+    assertThat(timelineLatch.await(TIMEOUT_MS, MILLISECONDS)).isTrue();
+    threadTestRule.getHandler().postAndSync(() -> controller.removeListener(timelineListener));
+    CountDownLatch latch = new CountDownLatch(2);
+    Player.Listener listener =
+        new Player.Listener() {
+          @Override
+          public void onAvailableCommandsChanged(Commands availableCommands) {
+            latch.countDown();
+          }
+
+          @Override
+          public void onTimelineChanged(Timeline timeline, int reason) {
+            capturedTimelines.add(timeline);
+            latch.countDown();
+          }
+        };
+    controller.addListener(listener);
+
+    remoteSession
+        .getMockPlayer()
+        .notifyAvailableCommandsChanged(
+            new Commands.Builder().addAllCommands().remove(COMMAND_GET_TIMELINE).build());
+
+    assertThat(latch.await(TIMEOUT_MS, MILLISECONDS)).isTrue();
+    assertThat(capturedTimelines).hasSize(2);
+    assertThat(capturedTimelines.get(0).getWindowCount()).isEqualTo(2);
+    assertThat(capturedTimelines.get(0).getWindow(0, new Timeline.Window()).isPlaceholder).isTrue();
+    assertThat(capturedTimelines.get(0).getWindow(1, new Timeline.Window()).isPlaceholder).isTrue();
+    assertThat(capturedTimelines.get(1).getWindowCount()).isEqualTo(1);
+    assertThat(capturedTimelines.get(1).getWindow(0, new Timeline.Window()).isPlaceholder)
+        .isFalse();
   }
 
   private void testControllerAfterSessionIsClosed(String id) throws Exception {